# openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: K8sPenTool-ng API
  description: |
    K8sPenTool-ng - Kubernetes Container Penetration Testing Platform

    Integrates CDK (Container Defense Kit) exploits with AI-driven attack chain automation.
    All endpoints use JSON request/response. Real-time events via WebSocket at `/api/v1/ws`.
  version: "2.0.0"
  contact:
    name: K8sPenTool Team
    url: https://github.com/trymonoly/K8sPenTool-ng
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /targets:
    post:
      tags: [Target]
      summary: Create a new target session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetRequest'
      responses:
        '201':
          description: Target created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
    get:
      tags: [Target]
      summary: List all targets
      responses:
        '200':
          description: List of targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Target'
  /targets/{id}:
    get:
      tags: [Target]
      summary: Get target details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Target details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
    delete:
      tags: [Target]
      summary: Delete a target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Target deleted
  /targets/{id}/steps:
    post:
      tags: [Target]
      summary: Record a step result for the active target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetStepRequest'
      responses:
        '200':
          description: Step recorded
  /proxy:
    get:
      tags: [Target]
      summary: Get global SOCKS5 proxy configuration
      responses:
        '200':
          description: Current proxy config or disabled status
    post:
      tags: [Target]
      summary: Set global SOCKS5 proxy configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyConfigRequest'
      responses:
        '200':
          description: Proxy configuration saved
    delete:
      tags: [Target]
      summary: Clear global SOCKS5 proxy configuration
      responses:
        '200':
          description: Proxy configuration cleared

  # Info
...
  /info/profiles/{id}/run:
    post:
      tags: [Info]
      summary: Run an evaluation profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            enum: [basic, extended, full]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequest'
      responses:
        '200':
          description: Evaluation started
  /info/port-scan:
    post:
      tags: [Info]
      summary: Scan K8s ports on target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host:
                  type: string
                ports:
                  type: string
              required: [host]
      responses:
        '200':
          description: Scan results
  /info/decode-capabilities:
    post:
      tags: [Info]
      summary: Decode Linux capability hex value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hex:
                  type: string
              required: [hex]
      responses:
        '200':
          description: Decoded capabilities
...
  /access/api-server:
    post:
      tags: [Access]
      summary: Check APIServer access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequest'
      responses:
        '200':
          description: Access check result
  /access/api-server/insecure:
    post:
      tags: [Access]
      summary: Check APIServer insecure port (8080)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequest'
      responses:
        '200':
          description: Port check result
  /access/api-server/request:
    post:
      tags: [Access]
      summary: Send custom request to APIServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomAPIRequest'
      responses:
        '200':
          description: API response

  # Access - Kubelet
  /access/kubelet:
    post:
      tags: [Access]
      summary: Check Kubelet (10250) access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequest'
      responses:
        '200':
          description: Access check result
  /access/kubelet/exec:
    post:
      tags: [Access]
      summary: Execute command via Kubelet API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubeletExecRequest'
      responses:
        '200':
          description: Command output

  # Access - Etcd
  /access/etcd/check:
    post:
      tags: [Access]
      summary: Check Etcd (2379) unauthorized access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                timeout_sec:
                  type: integer
              required: [target_host]
      responses:
        '200':
          description: Etcd access result
  /access/etcd/keys:
    post:
      tags: [Access]
      summary: List all Etcd keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Etcd keys
  /access/etcd/read:
    post:
      tags: [Access]
      summary: Read a specific Etcd key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                key:
                  type: string
              required: [target_host, key]
      responses:
        '200':
          description: Key value
  /access/etcd/v3/keys:
    post:
      tags: [Access]
      summary: Enumerate etcd v3 keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetHostRequest'
      responses:
        '200':
          description: Etcd v3 key range response
  /access/etcd/v3/search-secrets:
    post:
      tags: [Access]
      summary: Search etcd v3 for Kubernetes secret objects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetHostRequest'
      responses:
        '200':
          description: Etcd v3 secret search response
  /access/kubelet/inject-ssh:
    post:
      tags: [Access]
      summary: Try to inject an SSH public key via unauthenticated Kubelet exec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubeletSSHInjectRequest'
      responses:
        '200':
          description: Injection attempt results

  # Access - Dashboard
  /access/dashboard:
    post:
      tags: [Access]
      summary: Legacy dashboard probe alias
      description: Deprecated compatibility endpoint. Prefer `/dashboard/probe` for the unified Dashboard workflow.
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                port:
                  type: integer
                use_https:
                  type: boolean
              required: [target_host]
      responses:
        '200':
          description: Dashboard check result

  # Access - Kubeconfig
  /access/kubeconfig/parse:
    post:
      tags: [Access]
      summary: Parse kubeconfig content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        '200':
          description: Parsed kubeconfig info

  # Exec
  /exec/api-server/list-pods:
    post:
      tags: [Exec]
      summary: List pods via APIServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecListRequest'
      responses:
        '200':
          description: Pod list
  /exec/api-server/exec:
    post:
      tags: [Exec]
      summary: Execute command in pod via APIServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PodExecRequest'
      responses:
        '200':
          description: Command output
  /exec/api-server/enum-sa-tokens:
    post:
      tags: [Exec]
      summary: Enumerate service account tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                namespace:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: SA tokens
  /exec/kubelet/list-pods:
    post:
      tags: [Exec]
      summary: List pods via Kubelet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Pod list from Kubelet
  /exec/kubelet/exec:
    post:
      tags: [Exec]
      summary: Execute command via Kubelet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubeletExecRequest'
      responses:
        '200':
          description: Command output
  /exec/backdoor/generate-yaml:
    post:
      tags: [Exec]
      summary: Generate backdoor pod YAML
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackdoorConfig'
      responses:
        '200':
          description: Backdoor YAML
  /exec/rbac/check:
    post:
      tags: [Exec]
      summary: Check RBAC permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: RBAC check result
  /exec/reverse-shell/generate:
    post:
      tags: [Exec]
      summary: Generate reverse shell payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lhost:
                  type: string
                lport:
                  type: string
                type:
                  type: string
                  enum: [bash-i, bash, python, perl, nc, nc-e, nc-mkfifo, php, ruby, lua, curl]
              required: [lhost, lport]
      responses:
        '200':
          description: Reverse shell payload
  /exec/upload-file:
    post:
      tags: [Exec]
      summary: Upload a local file into a pod container
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadFileRequest'
      responses:
        '200':
          description: File upload result
  /exec/port-forward:
    post:
      tags: [Exec]
      summary: Generate port-forward and tunnel instructions for a pod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortForwardRequest'
      responses:
        '200':
          description: Port-forward guidance

  # Persist
  /persist/service-account:
    post:
      tags: [Persist]
      summary: Create high-privilege ServiceAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSARequest'
      responses:
        '200':
          description: SA creation result
  /persist/service-account/token:
    post:
      tags: [Persist]
      summary: Get ServiceAccount token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSARequest'
      responses:
        '200':
          description: SA token
  /persist/cronjob:
    post:
      tags: [Persist]
      summary: Deploy CronJob backdoor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CronJobRequest'
      responses:
        '200':
          description: CronJob deployment result
  /persist/daemonset:
    post:
      tags: [Persist]
      summary: Deploy DaemonSet backdoor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DaemonSetRequest'
      responses:
        '200':
          description: DaemonSet deployment result
  /persist/kubeconfig:
    post:
      tags: [Persist]
      summary: Generate shadow kubeconfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                server:
                  type: string
                cluster:
                  type: string
                token:
                  type: string
              required: [server, token]
      responses:
        '200':
          description: Generated kubeconfig
  /persist/host-crontab:
    post:
      tags: [Persist]
      summary: Generate host persistence commands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lhost:
                  type: string
                lport:
                  type: string
              required: [lhost, lport]
      responses:
        '200':
          description: Host persistence commands

  # Escape
...
  /escape/privileged:
    post:
      tags: [Escape]
      summary: Privileged container escape
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                namespace:
                  type: string
                pod_name:
                  type: string
                lhost:
                  type: string
                lport:
                  type: string
              required: [target_host, pod_name, lhost, lport]
      responses:
        '200':
          description: Escape commands
  /escape/mount:
    post:
      tags: [Escape]
      summary: Mount-based container escape
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                escape_type:
                  type: string
                  enum: [chroot, crontab, docker.sock, procfs]
                lhost:
                  type: string
                lport:
                  type: string
              required: [escape_type, lhost, lport]
      responses:
        '200':
          description: Escape commands
...
  /lateral/secrets:
    post:
      tags: [Lateral]
      summary: List secrets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                namespace:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Secret list
  /lateral/secrets/view:
    post:
      tags: [Lateral]
      summary: View and decode a secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                namespace:
                  type: string
                secret_name:
                  type: string
                token:
                  type: string
              required: [target_host, namespace, secret_name]
      responses:
        '200':
          description: Decoded secret data
  /lateral/services:
    post:
      tags: [Lateral]
      summary: List services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                namespace:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Service list
  /lateral/endpoints:
    post:
      tags: [Lateral]
      summary: List endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Endpoint list
  /lateral/nodes:
    post:
      tags: [Lateral]
      summary: List nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Node list
  /lateral/network-policies:
    post:
      tags: [Lateral]
      summary: List network policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Network policy list
  /lateral/taints:
    post:
      tags: [Lateral]
      summary: Show node taints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                token:
                  type: string
              required: [target_host]
      responses:
        '200':
          description: Node taints
  /lateral/taint-pod:
    post:
      tags: [Lateral]
      summary: Generate taint toleration pod YAML
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_host:
                  type: string
                namespace:
                  type: string
                node_name:
                  type: string
                image:
                  type: string
                host_mount:
                  type: boolean
              required: [target_host]
      responses:
        '200':
          description: Taint pod YAML

  # Kubectl
  /kubectl/exec:
    post:
      tags: [Kubectl]
      summary: Execute arbitrary kubectl command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlCommandRequest'
      responses:
        '200':
          description: Command output
  /kubectl/get-pods:
    post:
      tags: [Kubectl]
      summary: kubectl get pods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Pod list
  /kubectl/get-nodes:
    post:
      tags: [Kubectl]
      summary: kubectl get nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Node list
  /kubectl/get-services:
    post:
      tags: [Kubectl]
      summary: kubectl get services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Service list
  /kubectl/get-secrets:
    post:
      tags: [Kubectl]
      summary: kubectl get secrets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Secret list
  /kubectl/get-deployments:
    post:
      tags: [Kubectl]
      summary: kubectl get deployments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Deployment list
  /kubectl/cluster-info:
    post:
      tags: [Kubectl]
      summary: Query cluster version information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Cluster version details
  /kubectl/auth-can-i:
    post:
      tags: [Kubectl]
      summary: Evaluate effective RBAC permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Permission matrix and admin hint
  /kubectl/get-sa:
    post:
      tags: [Kubectl]
      summary: kubectl get serviceaccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Service account list
  /kubectl/get-crb:
    post:
      tags: [Kubectl]
      summary: kubectl get clusterrolebindings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: ClusterRoleBinding list
  /kubectl/get-images:
    post:
      tags: [Kubectl]
      summary: Aggregate images in running pods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlRequest'
      responses:
        '200':
          description: Image inventory
...
components:
  schemas:
...
    Target:
      type: object
      properties:
        id:
          type: string
        host:
          type: string
        port:
          type: integer
        token:
          type: string
        auth_type:
          type: string
          enum: [none, token, cert, kubeconfig, userpass]
        skip_tls:
          type: boolean
        timeout_sec:
          type: integer
        kubeconfig:
          type: string
        username:
          type: string
        password:
          type: string
    CreateTargetRequest:
      type: object
      required: [host]
      properties:
        host:
          type: string
        port:
          type: integer
          default: 6443
        token:
          type: string
        auth_type:
          type: string
          default: none
        skip_tls:
          type: boolean
          default: true
        timeout_sec:
          type: integer
          default: 10
        username:
          type: string
        password:
          type: string
        kubeconfig:
          type: string
...